moved loader service to shared folder
